chore: reorg file structures
