+++ packages/apidom-ns-workflows-1/src/refractor/specification.ts
@@ -11,9 +11,9 @@ import StepVisitor from './visitors/workflows-1/step';
 import StepOutputsVisitor from './visitors/workflows-1/step/OutputsVisitor';
 import StepParametersVisitor from './visitors/workflows-1/step/ParametersVisitor';
 import StepDependsOnVisitor from './visitors/workflows-1/step/DependsOnVisitor';
-import StepSuccessCriteriaVisitor from './visitors/workflows-1/StepSuccessCriteriaVisitor';
-import StepOnSuccessVisitor from './visitors/workflows-1/StepOnSuccessVisitor';
-import StepOnFailureVisitor from './visitors/workflows-1/StepOnFailureVisitor';
+import StepSuccessCriteriaVisitor from './visitors/workflows-1/step/SuccessCriteriaVisitor';
+import StepOnSuccessVisitor from './visitors/workflows-1/step/OnSuccessVisitor';
+import StepOnFailureVisitor from './visitors/workflows-1/step/OnFailureVisitor';
 import ParameterVisitor from './visitors/workflows-1/parameter';
 import SourceDescriptionsVisitor from './visitors/workflows-1/SourceDescriptionsVisitor';
 import SuccessActionVisitor from './visitors/workflows-1/success-action';
